Moves JsonComparator to datamodel.implementation package
