+++ wdtk-datamodel/src/test/java/org/wikidata/wdtk/datamodel/implementation/StringValueImplTest.java
@@ -25,7 +25,6 @@ import static org.junit.Assert.*;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Test;
-import org.wikidata.wdtk.datamodel.implementation.json.JsonComparator;
 import org.wikidata.wdtk.datamodel.interfaces.StringValue;
 
 import java.io.IOException;
