+++ wdtk-datamodel/src/test/java/org/wikidata/wdtk/datamodel/implementation/TermImplTest.java
@@ -23,7 +23,6 @@ package org.wikidata.wdtk.datamodel.implementation;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Test;
-import org.wikidata.wdtk.datamodel.implementation.json.JsonComparator;
 import org.wikidata.wdtk.datamodel.interfaces.MonolingualTextValue;
 
 import java.io.IOException;
